refactor: extract session lifecycle into session-lifecycle.ts
Move Claude process lifecycle methods (startClaude, waitForReady,
wireClaudeEvents, handleClaudeExit, stopClaude, stopClaudeAndWait)
from SessionManager into a dedicated SessionLifecycle class.

SessionManager retains thin delegation wrappers so the public API
is unchanged. Follows the same dependency-injection pattern used
by PromptRouter (PR #303).
